add hall feedback motor run function

+++ tools.h
@@ -3,9 +3,9 @@
 
 #include <xc.h> // include processor files - each processor file is guarded.  
 
-static inline uint8_t Scaling100(const uint16_t value, uint16_t value_max, uint16_t target_max) 
+static inline uint16_t Scaling100(const uint16_t value, uint32_t value_max, uint32_t target_max) 
 {
-    return (uint16_t)((uint32_t)(value * target_max) / value_max);
+    return (uint16_t)((value * target_max) / value_max);
 }
 
 #endif /* TOOLS_H */
